requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.1

# Document parsing

PyMuPDF>=1.24.9
python-docx>=1.1.0

# LLMs

openai>=1.40.0

# Utilities

requests>=2.32.0
python-dotenv>=1.0.0
